Crate gitpad

Re-exports

pub use data::Data;

Modules

api 🔒

data

db 🔒

demo

Demo user: Enable users to try out your application without signing up

errors

represents all the ways a trait can fail using this crate

pages 🔒

routes 🔒

settings 🔒

static_assets 🔒

utils 🔒

Structs

FILES

Settings

Constants

CACHE_AGE

GIT_COMMIT_HASH

PKG_DESCRIPTION

PKG_HOMEPAGE

PKG_NAME

V1_API_ROUTES

constant Routes instance

VERSION

Functions

get_identity_service

get_json_err

main 🔒

Type Definitions

AppData

DB